+++ ignite/cmd/generate_dart.go
@@ -4,6 +4,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/ignite/cli/ignite/pkg/cliui"
+	"github.com/ignite/cli/ignite/pkg/cliui/icons"
 	"github.com/ignite/cli/ignite/services/chain"
 )
 
@@ -21,16 +22,15 @@ func NewGenerateDart() *cobra.Command {
 }
 
 func generateDartHandler(cmd *cobra.Command, args []string) error {
-	session := cliui.New(cliui.StartSpinner())
+	session := cliui.New(cliui.StartSpinnerWithText(statusGenerating))
 	defer session.End()
 
-	session.StartSpinner("Generating...")
-
-	c, err := newChainWithHomeFlags(
+	c, err := NewChainWithHomeFlags(
 		cmd,
 		chain.EnableThirdPartyModuleCodegen(),
 		chain.WithOutputer(session),
 		chain.CollectEvents(session.EventBus()),
+		chain.PrintGeneratedPaths(),
 	)
 	if err != nil {
 		return err
@@ -45,5 +45,5 @@ func generateDartHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	return session.Println("⛏️  Generated Dart client.")
+	return session.Println(icons.OK, "Generated Dart Client")
 }