+++ ignite/cmd/generate_typescript_client.go
@@ -4,6 +4,7 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/ignite/cli/ignite/pkg/cliui"
+	"github.com/ignite/cli/ignite/pkg/cliui/icons"
 	"github.com/ignite/cli/ignite/services/chain"
 )
 
@@ -22,16 +23,15 @@ func NewGenerateTSClient() *cobra.Command {
 }
 
 func generateTSClientHandler(cmd *cobra.Command, args []string) error {
-	session := cliui.New(cliui.StartSpinner())
+	session := cliui.New(cliui.StartSpinnerWithText(statusGenerating))
 	defer session.End()
 
-	session.StartSpinner("Generating...")
-
-	c, err := newChainWithHomeFlags(
+	c, err := NewChainWithHomeFlags(
 		cmd,
 		chain.EnableThirdPartyModuleCodegen(),
 		chain.WithOutputer(session),
 		chain.CollectEvents(session.EventBus()),
+		chain.PrintGeneratedPaths(),
 	)
 	if err != nil {
 		return err
@@ -52,5 +52,5 @@ func generateTSClientHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	return session.Println("⛏️  Generated Typescript Client")
+	return session.Println(icons.OK, "Generated Typescript Client")
 }