+++ ignite/services/network/join.go
@@ -2,13 +2,11 @@ package network
 
 import (
 	"context"
-	"fmt"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	launchtypes "github.com/tendermint/spn/x/launch/types"
 
 	"github.com/ignite/cli/ignite/pkg/cosmosutil"
-	"github.com/ignite/cli/ignite/pkg/events"
 	"github.com/ignite/cli/ignite/pkg/xurl"
 	"github.com/ignite/cli/ignite/services/network/networkchain"
 	"github.com/ignite/cli/ignite/services/network/networktypes"
@@ -85,60 +83,19 @@ func (n Network) Join(
 	}
 
 	if !o.accountAmount.IsZero() {
-		if err := n.sendAccountRequest(ctx, launchID, accountAddress, o.accountAmount); err != nil {
+		if err := n.SendAccountRequest(ctx, launchID, accountAddress, o.accountAmount); err != nil {
 			return err
 		}
 	}
 
-	return n.sendValidatorRequest(ctx, launchID, peer, accountAddress, gentx, gentxInfo)
+	return n.SendValidatorRequest(ctx, launchID, peer, accountAddress, gentx, gentxInfo)
 }
 
-// sendValidatorRequest creates the RequestAddValidator message into the SPN
-func (n Network) sendValidatorRequest(
-	ctx context.Context,
-	launchID uint64,
-	peer launchtypes.Peer,
-	valAddress string,
-	gentx []byte,
-	gentxInfo cosmosutil.GentxInfo,
-) error {
+func (n Network) SendAccountRequestForCoordinator(ctx context.Context, launchID uint64, amount sdk.Coins) error {
 	addr, err := n.account.Address(networktypes.SPN)
 	if err != nil {
 		return err
 	}
 
-	msg := launchtypes.NewMsgSendRequest(
-		addr,
-		launchID,
-		launchtypes.NewGenesisValidator(
-			launchID,
-			valAddress,
-			gentx,
-			gentxInfo.PubKey,
-			gentxInfo.SelfDelegation,
-			peer,
-		),
-	)
-
-	n.ev.Send("Broadcasting validator transaction", events.ProgressStarted())
-
-	res, err := n.cosmos.BroadcastTx(ctx, n.account, msg)
-	if err != nil {
-		return err
-	}
-
-	var requestRes launchtypes.MsgSendRequestResponse
-	if err := res.Decode(&requestRes); err != nil {
-		return err
-	}
-
-	if requestRes.AutoApproved {
-		n.ev.Send("Validator added to the network by the coordinator!", events.ProgressFinished())
-	} else {
-		n.ev.Send(
-			fmt.Sprintf("Request %d to join the network as a validator has been submitted!", requestRes.RequestID),
-			events.ProgressFinished(),
-		)
-	}
-	return nil
+	return n.SendAccountRequest(ctx, launchID, addr, amount)
 }