+++ changelog.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- [#2955](https://github.com/ignite/cli/pull/2955/) Add `ignite network request add-account` command.
+- [#2877](https://github.com/ignite/cli/pull/2877) Plugin system
+- [#2995](https://github.com/ignite/cli/pull/2995/) Add `ignite network request remove-validator` command.
+- [#2999](https://github.com/ignite/cli/pull/2999/) Add `ignite network request remove-account` command.
+
+### Changes
+
+- [#2957](https://github.com/ignite/cli/pull/2957) Change generate commands to print the path to the generated code.
+- [#2981](https://github.com/ignite/cli/issues/2981) Change CLI to also search chain binary in Go binary path.
+- [#2958](https://github.com/ignite/cli/pull/2958) Support absolute paths for client code generation config paths.
+- [#2993](https://github.com/ignite/cli/pull/2993) Hide `ignite scaffold band` command and deprecate functionality.
+- [#2991](https://github.com/ignite/cli/pull/2991) Hide `ignite scaffold flutter` command and remove functionality.
+- [#2944](https://github.com/ignite/cli/pull/2944) Add a new event "update" status option to `pkg/cliui`.
+
+## [`v0.25.1`](https://github.com/ignite/cli/releases/tag/v0.25.1)
+
+### Changes
+
+- [#2968](https://github.com/ignite/cli/pull/2968) Dragonberry security fix upgrading Cosmos SDK to `v0.46.3`
+
 ## [`v0.25.0`](https://github.com/ignite/cli/releases/tag/v0.25.0)
 
 ### Features
@@ -27,7 +51,7 @@
 - Add `--peer-address` flag to `network chain join` command.
 - Change nightly tag format
 - Add cosmos-sdk version in `version` command
-- [#29350](https://github.com/ignite/cli/pull/2935) Update `gobuffalo/plush` templating tool to `v4`
+- [#2935](https://github.com/ignite/cli/pull/2935) Update `gobuffalo/plush` templating tool to `v4`
 
 ### Fixes